+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Do not include archived assets in metrics count.
 - Remove "administer farm map" from Manager role permissions.
 - [Add allow-plugins config #467](https://github.com/farmOS/farmOS/pull/467)